add: Companyリソースのアクセス制限を追加#26

+++ server/app/controllers/api/v1/companies_controller.rb
@@ -45,6 +45,7 @@ module Api
 
       # 新規作成
       def create
+        authorize Company
         company = current_api_v1_user.companies.new(company_params)
         if company.save
           render json: company
@@ -60,6 +61,7 @@ module Api
 
       # 会社情報更新
       def update
+        authorize Company
         if @company.update(company_params)
           render json: @company
         else
@@ -69,6 +71,7 @@ module Api
 
       # 会社情報削除
       def destroy
+        authorize Company
         if @company.destroy
           render json: @company
         else
